+++ NEWS.md
@@ -1,3 +1,3 @@
-# checkR 0.0.0.9000
-
-* Added a `NEWS.md` file to track changes to the package.
+# checkR (development version)
+* added first checks
+* Initial CRAN submission.